#77 Refactor: 패키지 구조를 개선한다.
