+++ src/main/java/org/zapply/product/domain/posting/service/PostingQueryService.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.zapply.product.domain.user.entity.Member;
 import org.zapply.product.domain.user.service.AccountService;
 import org.zapply.product.global.clova.enuermerate.SNSType;
-import org.zapply.product.global.threads.ThreadsMediaClient;
-import org.zapply.product.global.threads.ThreadsMediaResponse;
+import org.zapply.product.global.snsClients.threads.ThreadsMediaClient;
+import org.zapply.product.global.snsClients.threads.ThreadsMediaResponse;
 
 import java.util.List;
