+++ src/main/java/org/zapply/product/domain/user/service/AccountService.java
@@ -13,15 +13,15 @@ import org.zapply.product.domain.user.repository.AccountRepository;
 import org.zapply.product.domain.user.repository.MemberRepository;
 import org.zapply.product.global.apiPayload.exception.CoreException;
 import org.zapply.product.global.apiPayload.exception.GlobalErrorType;
-import org.zapply.product.global.facebook.FacebookClient;
-import org.zapply.product.global.facebook.FacebookProfile;
-import org.zapply.product.global.facebook.FacebookToken;
-import org.zapply.product.global.linkedin.LinkedinClient;
-import org.zapply.product.global.linkedin.LinkedinToken;
-import org.zapply.product.global.linkedin.LinkedinUserInfo;
-import org.zapply.product.global.threads.ThreadsClient;
-import org.zapply.product.global.threads.ThreadsProfile;
-import org.zapply.product.global.threads.ThreadsToken;
+import org.zapply.product.global.snsClients.facebook.FacebookClient;
+import org.zapply.product.global.snsClients.facebook.FacebookProfile;
+import org.zapply.product.global.snsClients.facebook.FacebookToken;
+import org.zapply.product.global.snsClients.linkedin.LinkedinClient;
+import org.zapply.product.global.snsClients.linkedin.LinkedinToken;
+import org.zapply.product.global.snsClients.linkedin.LinkedinUserInfo;
+import org.zapply.product.global.snsClients.threads.ThreadsClient;
+import org.zapply.product.global.snsClients.threads.ThreadsProfile;
+import org.zapply.product.global.snsClients.threads.ThreadsToken;
 import org.zapply.product.global.vault.VaultClient;
 
 import java.security.MessageDigest;
