+++ src/main/java/org/zapply/product/global/snsClients/facebook/FacebookProfile.java
@@ -1,4 +1,4 @@
-package org.zapply.product.global.facebook;
+package org.zapply.product.global.snsClients.facebook;
 
 public record FacebookProfile(
         String id,
