+++ src/main/java/org/zapply/product/global/snsClients/facebook/FacebookToken.java
@@ -1,4 +1,4 @@
-package org.zapply.product.global.facebook;
+package org.zapply.product.global.snsClients.facebook;
 
 public record FacebookToken(
         String accessToken
