+++ src/main/java/org/zapply/product/global/snsClients/threads/ThreadsReplyClient.java
@@ -1,0 +1,4 @@
+package org.zapply.product.global.snsClients.threads;
+
+public class ThreadsReplyClient {
+}
