--- src/main/java/org/zapply/product/global/threads/ThreadsReplyClient.java
@@ -1,4 +1,0 @@
-package org.zapply.product.global.threads;
-
-public class ThreadsReplyClient {
-}
